案件手工分配

+++ service/business/src/main/java/cn/fintecher/pangolin/business/model/AllocationCountModel.java
@@ -4,6 +4,9 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+
 /**
  * Created by sunyanping on 2017/9/26.
  */
@@ -11,7 +14,7 @@ import lombok.Data;
 @ApiModel(description = "手动分案案件统计")
 public class AllocationCountModel {
     @ApiModelProperty("案件总个数")
-    private Integer caseTotal;
+    private BigInteger caseTotal;
     @ApiModelProperty("案件总金额")
-    private Double caseAmount;
+    private BigDecimal caseAmount;
 }
